+++ Runtime/Inputs/Simulation/XRSimulatedController.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine.InputSystem.Controls;
 using UnityEngine.InputSystem.Layouts;
+using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.Scripting;
 
 namespace UnityEngine.XR.Interaction.Toolkit.Inputs.Simulation
@@ -119,6 +120,14 @@ namespace UnityEngine.XR.Interaction.Toolkit.Inputs.Simulation
             secondary2DAxisTouch = GetChildControl<ButtonControl>(nameof(secondary2DAxisTouch));
             batteryLevel = GetChildControl<AxisControl>(nameof(batteryLevel));
             userPresence = GetChildControl<ButtonControl>(nameof(userPresence));
+        }
+
+        /// <inheritdoc />
+        protected override unsafe long ExecuteCommand(InputDeviceCommand* commandPtr)
+        {
+            return XRDeviceSimulator.TryExecuteCommand(commandPtr, out var result)
+                ? result
+                : base.ExecuteCommand(commandPtr);
         }
     }
 }