+++ Tests/Editor/InteractionLayerUpdaterTest.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 using UnityEngine.XR.Interaction.Toolkit;
 using UnityEngine.XR.Interaction.Toolkit.Tests;
 
-namespace UnityEditor.XR.Interaction.Toolkit.Tests
+namespace UnityEditor.XR.Interaction.Toolkit.Editor.Tests
 {
     [TestFixture]
     class InteractionLayerUpdaterTest